feat: parsing html for rfc link previews etc
